+++ src/LivewireTableKitServiceProvider.php
@@ -13,6 +13,7 @@ class LivewireTableKitServiceProvider extends ServiceProvider
         if ($this->app->runningInConsole()) {
             $this->commands([
                 \Unlab\LivewireTableKit\Console\Commands\MakeTableCommand::class,
+                \Unlab\LivewireTableKit\Console\Commands\InstallSkillCommand::class,
                 \Unlab\LivewireTableKit\Console\Commands\McpServerCommand::class,
                 \Unlab\LivewireTableKit\Console\Commands\InstallMcpCommand::class,
             ]);